Refactor LoginPage
